% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Clean strings}
\usage{
clean_string(x)
}
\arguments{
\item{x}{a vector of characters. Typically a vector of titles/abstracts to
clean.}
}
\description{
Cleans a vector of strings by removing multiple whitespace, number,
punctuation, etc.
}
