% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\title{Screen references title/abstract}
\usage{
screen(x, pattern)
}
\arguments{
\item{x}{a vector of characters. Typically a vector of titles/abstracts in
which \code{pattern} must be detected.}

\item{pattern}{a vector of characters. Patterns to detect in \code{x}.}
}
\description{
Screens references title/abstract. Detect patterns (argument \code{pattern}) in a
vector of strings (argument \code{x}).
}
