% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenization}
\alias{tokenization}
\title{Combine expressions in single word}
\usage{
tokenization(x, token)
}
\arguments{
\item{x}{a vector of characters. Typically a vector of titles/abstracts in
which expressions listed in \code{token} must be replaced.}

\item{token}{a vector of characters. Words expression to convert in single
words}
}
\description{
Combines expressions in single word. For instance, \verb{climate change} will
become \code{climatechange}.
}
\examples{
x <- "funrar: An R package to characterize functional rarity"
tokenization(x, token = c("R package", "functional rarity"))
}
